clang test

+++ build.py
@@ -8,11 +8,19 @@ import Cmake
 import Shell
 import Compiler
 
+os.environ['CC'] = 'clang'
+os.environ['CXX'] = 'clang++'
+
+print(os.environ['CXX'])
+
 Conan.setup()
 
 build_folder = File.build_folder()
 
-File.rm(build_folder)
+
+if Args.rmbuild:
+	File.rm(build_folder)
+	
 File.mkdir(build_folder)
 File.cd(build_folder)
